retention look up by exact prev_ts, checking ts against retention (#34449)
Redo of #34317 . Instead of querying in the documents log for exactly `(id, prev_ts)`, we change the interface to take in an `id, ts, prev_ts`. Then we can compare the `ts` against retention and use `prev_ts` for the direct equality lookup. And we can return a map that uses the entire `id, ts, prev_ts` as a key, to avoid ambiguity of which timestamp is used in the key.

GitOrigin-RevId: 1e8ac19ad499ad4751820b92cc2580b9bc433c78