create _canonical_urls system table (#34431)
create a new system table `_canonical_urls` which will be the source of truth for the canonical urls of an instance.
This data is related to but stored independently from the custom/vanity domains, which are used at the routing layer.
In contrast, the canonical urls are owned by the backend/instance, so they can be used in `process.env.CONVEX_CLOUD_URL` and `process.env.CONVEX_SITE_URL`.

Note the storage of urls instead of domains, since the urls can potentially be `http://` instead of `https://`, or have a `/http` path at the end.

The overall plan is for the backend to own this "canonical url" information, and big brain doesn't care if those get out of sync with the verified custom domains (that big brain owns). These canonical urls will be used as the system env variables and also will be publically accessible so they can be used for `VITE_CONVEX_URL` by the CLI and such.

The table is empty by default, in which case the urls used are the .convex.cloud and .convex.site ones.

GitOrigin-RevId: aaa0bc6e0824d10f47d09c45bbc141d99d771d17