new system-udf to query `process.env.CONVEX_CLOUD_URL` (#34464)
create a new system-udf that's a query returning `process.env.CONVEX_CLOUD_URL`. this is currently redundant but will be used once we add canonical custom domains.

It'll be used by the dashboard, which displays the canonical domain and lets you pick a new one.

It will also be used by the CLI, which will set it to `VITE_CONVEX_URL` during deployment.

Put it in a file in `_system/cli/` which indicates it has a higher standard of backwards-compatibility, but also re-export it from `_system/frontend/` to indicate it's used in the dashboard too.

(for review to @jordanhunt22 for canonical urls and @atrakh for system udf placement and naming)

GitOrigin-RevId: a7c278a18e941dac8a6ede9cdeb2936667606bec