use process.env.CONVEX_CLOUD_URL for storage getUrl (#34729)
when calling ctx.storage.getUrl(), use the canonical cloud url, if any, as the base.

add a test.

note this doesn't change generateUploadUrl, which still uses the .convex.cloud url as the base.

an alternative design would be to set the canonical url as the origin when constructing TransactionalFileStorage. That has the benefit of also handling generateUploadUrl, but it becomes harder to make it reactive and in some cases like node actions would use old values from when conductor restarted. The most obvious way to make it reactive and consistent is to read it with the active transaction. I'll figure out generateUploadUrl later.

GitOrigin-RevId: 2860be2dfed861e465f1b71e7c41ed482e3c1ce9