storage.generateUploadUrl use process.env.CONVEX_CLOUD_URL (#34733)
Use canonical urls when constructing storage upload urls.
For most things it's straightforward because we have a transaction or can construct one easily, following the pattern of `storage.getUrl`.

The tricky one was the v8 action async syscall. The pattern from `storage.getUrl` is to use an action callback, but that seems like a lot of additional code and overhead for something that should be simple. Fetching it with the rest of the action prefetching seems to follow a different existing pattern and works.

GitOrigin-RevId: 08381b8d6525a87854cd21f460437b2798dc2430