convex-backend PR 51: Add second workaround to circular type inference (#34901)
I added a second option to solve the circular type inference in actions: typing the handler function return type. It might be simpler than typing `runQuery` and `runMutation`, specially when there are multiple queries and mutations, so I place it first.

I also made the title an h3, to be visible in the TOC, and because I think it is more of a problem of the action constructor been part of the type generation than the context. Maybe it need to be part of another page (I don't known witch could be), because this could also happen on other functions, like mutations with partial rollback 🤔

Co-authored-by: Benjamín Vicente <benjavicente@uc.cl>
GitOrigin-RevId: 1e3981bfd62823d5d70bc0ef3c87a2968bfc323b