display canonical deployment url in dashboard (#34904)
we show the deployment url in the dashboard, which can be used several things, like connecting to airbyte.

this is currently the .convex.cloud url, but we can show the canonical process.env.CONVEX_CLOUD_URL instead.

note there's currently a loading state where we show the .convex.cloud url, which matches the similar state for http actions where we transiently show .convex.site. I'm not sure if we want to keep this loading state or just say "loading..." to make sure we don't show the wrong one. The .convex.site and .convex.cloud urls still work, but they might not be what the user wants or expects.

GitOrigin-RevId: 4775cf7417835451c8cda8299eefb3c6d13c4076