use TableIterator to calculate file storage usage (#34980)
Currently file storage is counted in a large transaction.

This happens to work today, because we don't have limits on the number of system documents or bytes read in a query. We have limits on basically everything else -- user documents and bytes, writes of system documents, number of readset intervals. But this particular gap in the limits allows a query to tablescan `_file_storage`.

Eventually a file storage table will get so big that this query cannot finish within the index retention window.

So let's remove this time bomb, by using TableIterator instead. We only need to do it to calculate usage from a background worker, so the additional latency shouldn't be an issue.

GitOrigin-RevId: a2fb1d14a9ce7837ea1219729d7aef95a13a304f