make _creationTime required (#35077)
this field has existed for a long time on all documents, and document retention has removed the documents log entries that don't have creation time. So we can make it required.

Note we can always make the field opt-out later, for product reasons. But for now we can make the code match what actually exists, so there's no confusion about scenarios where we might have to handle missing creation times. On the other hand, if we keep it optional, it makes sure we never design a feature that absolutely requires creation time. But I don't really see that being a benefit -- we already implicitly require creation times, like in `SystemTableCleanupWorker` which would get stuck in an infinite loop if there were no creation times.

GitOrigin-RevId: 779aec848a0ae298067a49ea370b52a0dd5ce552