dashboard: update paginatedTableDocuments to support disabling filters (#35061)
Adds backwards-compatible support for an `enabled` field on filters for the data page. Fields that set this field to false will not be including in filtering.

Also adds a test configuration for `system-udfs` because we moved tests there at some point, but did not make sure they were still running.

GitOrigin-RevId: 85a53d643e2c31d57a2686bd721a7730e4f204ea