fix site url picker in dev (#34941)
This is tricky because the backend doesn't expose its default http actions url. So when we see that the current http actions url is http://127.0.0.1:4532, the dashboard doesn't know whether this is the default, so it doesn't know whether it should show an option for changing the url back to the default, and what such an option would be labeled.

however, we can figure out a heuristic that should usually work:

1. if the current cloud url ends with .convex.cloud, then the http actions url is the same but with .convex.site. this covers the prod case.
2. if the current http actions url is not a registered custom domain, it is the default http actions url. this covers the initial dev case, since http://127.0.0.1:4583 won't be a registered custom domain.
3. if the current http actions url *is* a registered custom domain, the dashboard doesn't know what the default url should be. clicking "default" will select it. this covers the subsequent dev case, allowing you to deselect a custom url.

<img width="1409" alt="Screenshot 2025-03-03 at 9 16 37 PM" src="https://github.com/user-attachments/assets/2b9b3ebc-744f-46d1-9aab-3ff2cbbd9aad" />
<img width="1409" alt="Screenshot 2025-03-03 at 9 16 41 PM" src="https://github.com/user-attachments/assets/0a71acc0-6822-4c97-bdfb-c0b55185587e" />

GitOrigin-RevId: 169b1ba76d4ae8006eb4d00fecf18bda840472e3