convex-backend PR 59: Improve `.unique()` error message: print document `_id`s (#35342)
Makes the error message thrown by `.unique()` more useful. The `_id`'s of the first two documents returned by the `unique()` query are printed in the error message. This is very helpful to the user in identifying and resolving incorrect data.

Feel free to change my wording in the error message - but in my opinion it should be clear that the document id's printed are *not* all the documents returned by the `unique()` query but just the first two.

I [built the Convex backend locally](https://github.com/get-convex/convex-backend/blob/ad552652f4aaa0225e977ed7f5d92a35108b85d4/BUILD.md), made this change locally, and ran the tutorial demo to verify that it works as expected.

----

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.

Co-authored-by: Jan Balin <jgbalin@gmail.com>
GitOrigin-RevId: 816362f3274acabe1440eda6c6d993a94160009e