remove DISABLE_FUZZY_TEXT_SEARCH knob (#35357)
Removes the `DISABLE_FUZZY_TEXT_SEARCH` knob since it has been enabled for the past two months. This really just scratches the surface of the dead code we can remove. Will follow up eventually to get rid of the rest. Notably, this preserves the field in protos, but after this is pushed and we no longer read it, we can remove the proto field.

GitOrigin-RevId: 2b8c1f9f4b733e3bea3525bdba5049835f5d40f1