Keep error around when http actions auth fails (#35277)
Keep error around when http actions auth fails. Rethrow it only
when they actually call getUserIdentity. Will show up in the logs
serverside. Is a 500 to the client.

For now - keep the old behavior and log a metric whenever it happens - so we can see how many folks are affected in practice. Then we can follow up to turn the behavior on.

Motivated by https://github.com/get-convex/convex-backend/issues/58

GitOrigin-RevId: 711b78076ba6101e41dc7e3a60eaf236cf1459a1