Allow AbortSignal to cancel fetch (#35197)
Support `signal` in V8 fetch and Request, that makes the fetch get canceled.
It's implemented as a ReadableStream which gets closed when the signal fires. On the rust side, the ReadableStream becomes a Stream which we turn into a Future by waiting for it to close.

Note it throws a regular BadRequest, not a DOMException like the spec says. But it does work.

Tested manually in an action. Tried to test in an isolate test with pause controller, but pause is best supported with TestRuntime while `fetch` isn't supported in TestRuntime at all.

The request passed to HTTP actions now has a noop signal, at least until i implement signals for http action disconnects.

GitOrigin-RevId: a32c17a9b27c9e0c12645b7f73b1ea1147e64aa5