Stop depending on sodiumoxide (#35449)
The crate is unmaintained and we only need it for its secretbox primitive.
We can get that from `libsodium-sys-stable` (which is _not_ unmaintained) and write our own tiny binding.

Also get rid of the explicit `init()` at the top of every service, and just call init before invoking any libsodium routines. This is fast enough and the only downside is that an init error won't be caught immediately, but an init error cannot happen anyway.

GitOrigin-RevId: ed3f061302206ffd4024a2a2ec4106bd0b2f8ade