dashboard: add checks to big brain for network failures (#35428)
Adds checks to the big brain request loop that stops us from revalidating or retrying queries if it appears that big brain is offline. Instead, we'll hit the `/version` endpoint of big brain with exponential backoff until we get a successful response. Once we get a successful response, we'll set the app back to an online state

GitOrigin-RevId: c8a6d6c78503e2090c3390dc3b3c7a897bf55f90