Make SubscriptionsClient::subscribe synchronous (#35529)
There is no need to wait for the subscription worker to catch up before issuing a new subscription.
We can create the `Subscription` object immediately and 'attach' it to the subscription manager at a later time.
This is OK because 1. the token is refreshed (synchronously) at creation time anyway, via the WriteLog; and 2.
the subscription's `current_ts` is never a lie, it could just be lagging - but a caller that needs a hard guarantee
must call `extend_validity` with a desired timestamp anyway.

This PR also changes how the worker detects dropped subscriptions, by watching for `valid_tx.closed()` in
a FuturesUnordered rather than sending an explicit close message on Drop. This fixes an edge case during
cancellation where a subscription could get leaked if the subscription worker's queue was full when the
`Subscription` was dropped, leading to the Close message being forgotten.

GitOrigin-RevId: 8bf54815a1679a2b6cee7bb88880e595e66b4717