Refactor the CLI around deployment selection (#35207)
The goal of this refactor is to centralize all the logic around figuring out which deployment to talk to + clean up our environment variable handling. I want to do this before adding more flows so it's easier to reason about what's happening.

Each of our commands goes through 3 main steps:
* We create the `Context` which sets up the auth header we use for big brain
* We figure out which project we want to target, usually from the `CONVEX_DEPLOYMENT` env var. There's a fast path for "I know exactly which deployment I want to talk to" (e.g. self hosting, `--url` + `--admin-key`), and there's also an option for "chooseProject" -- I call this `deploymentSelection`
* Most commands will combine this project information with any flags like `--prod` or `--deployment-name` you passed in (I renamed this to `selectionWithinProject`), and give you the credentials of a deployment in that project. After that, the command is basically ready to go.

`npx convex dev` is special and supports creation of new projects. `npx convex deploy` is special and supports special behavior when it has a preview deploy key

This eliminates places where we were reading directly from `process.env` / mutating it directly and attempts to centralize more of the logic on all the env vars that can be set and which ones are incompatible.

GitOrigin-RevId: c1757ed3fe5b1de25980a40088b7a066f13722dc