fix http action warning (#35663)
See here: https://github.com/hyperium/hyper/blob/68bae2b53f8db4b0440e49b7e151e7a99229146c/src/proto/h1/decode.rs#L153

If an http action returns a Content-Length header, hyper/axum will stop polling the stream when it receives that many bytes. This is a problem for our body streams which do cleanup after yielding the last chunk. In a race condition, we are seeing:

1. we yield the last chunk, so that we have returned Content-Length bytes
2. hyper drops the stream
3. the drop propagates through usher to conductor, where it gets detected in the spawned thread which hasn't been dropped https://github.com/get-convex/convex/blob/cc2d57ad2b67b6ec4d7f495bbb7825d4afd2e190/crates/application/src/lib.rs#L1270
4. conductor cancels the funrun request and logs a warning to the execution log

to fix this, we should detect that we're about to return 'Content-Length' bytes and poll to the next value of the stream, which is probably None because the stream is closing.

tested by adding a sleep after yielding each chunk, so the cleanup is delayed until after hyper has dropped the stream. before this PR, that caused the warning every time. after this PR, there's no warning.

considered testing via pause controller, but it sounds really hard to have a pause in coductor while testing the hyper/axum/usher code.

GitOrigin-RevId: a050596953d9bf3f5dca83544300bff4b42b5a2c