Only compute snapshot updates once in committer (#35593)
This PR improves committer performance by removing calculating the new snapshot in `publish_commit`. Instead, we save the snapshot we already calculated in `validate_commit` and calculate the snapshot from the latest pending writes snapshot. We are guaranteed that the pending writes are applied in order, so this is safe.

GitOrigin-RevId: 99f843cbf2dd4fb0ce0f27ec53b853467cbd3043