[http actions] peek before response if empty (#35698)
Follow up to #35663 .

There's another problematic case: if the returned 'Content-Length' is zero, hyper/axum might not poll the body stream at all. That means we won't be able to do cleanup and we might get "Content disconnect" warnings in logs.

To fix this case, if Content-Length is known to be zero, we peek at the body -- presumably getting a None -- before returning any response at all.

GitOrigin-RevId: 8cca04da082652430fac8d1d32adc04dafbd001b