http actions keep running after client disconnects (#35689)
effective revert of https://github.com/get-convex/convex/pull/24942 for http actions.
when the client goes away we were cancelling the http action, but instead we should let it keep running to completion, while triggering the `request.signal` so it can stop and clean up if it wants to.

added test that fails before this change.

GitOrigin-RevId: 44af8b2bdb17489c01d9448109f8ef3d468b527a