Fix issues with project keys (#35726)
The previous code was mistakenly always returning "does not have access to project" instead of "cannot determine if has access" for project keys.

Additionally, we'd prompt for opt ins with preview or project keys, which wouldn't work. Also we'd try and check usage which would fail with the project key.

GitOrigin-RevId: 5079ce3b28d4d6686de351722b1e27789ac5cfb4