Lazily initialize LocalNodeExecutor (#35805)
We use the `LocalNodeExecutor` for folks who are self-hosting (e.g. using docker) as well as folks who are running the backend locally for development via `npx convex dev`.

In the latter case, developers might have a different node version installed than the one we normally use for node actions. Previously, we would detect the mismatched node version on `invoke` and throw a user visible error encouraging the developer to use `nvm` or similar to run Node 18.

When we switched to running a persistent node server instead, we moved this error to happen for everyone (not just folks running node actions) and also happen at start up where we can't easily show the error message to the developer (compared to errors hit during push or when running a function).

This starts the node server on the first `invoke` using `async_once_cell`. An alternative would've been using `enable` to kick off starting the node server but (1) it would require a separate worker to drive the future and (2) if it errored, we wouldn't be able to easily surface the error message to the developer.

GitOrigin-RevId: 7251af1fcbd023fe7faef86d00b887c2aa276c18