Add initial implementation of tryitout deployments (#35766)
This is a new flow where folks don't need to be logged in to use Convex called "tryitout" for now.

It's gated behind `CONVEX_TRY_IT_OUT=1`, so this code should be safe to land as is. It's also available for manual testing at convex@1.21.1-alpha.1

- If you're not logged in with no `CONVEX_DEPLOYMENT` set, prompt through choosing or creating a `tryitout` deployment
- If you’re not logged in and have a `tryitout` deployment in `CONVEX_DEPLOYMENT`, use that one
- If you’re not logged in and have a non `tryitout` deployment ⇒ print a message about how they can try Convex without logging in if they clear `CONVEX_DEPLOYMENT` and prompt for log in
- If you’re logged in and have a `tryitout` deployment ⇒ print message on how to link your deployment to a project and prompt to instead select a project associated with this member.
- Existing flows

This leverages a lot of the same code for local deployments.

This downloads the dashboard (at the same version as the local backend) and runs it locally. It also runs a "mini big brain" on an adjacent port that has a single endpoint that returns the deployment names, URLs, and admin keys for all the "tryitout" deployments.

There will be a command to migrate "tryitout" deployments to be local dev deployments for each project in your account that is not included in this PR.

GitOrigin-RevId: 61404c3ef1718be7cbc8e80d655f036dae8bed4e