Add ability to link tryitout deployments on login (#35877)
If someone tries Convex without an account and later decides to create an account, they can run `npx convex login` which will prompt the developer to link their "tryitout" deployments with their new account.

By default, for each "tryitout" deployment, we'll create a new project and make the "tryitout" deployment the local dev deployment for the new project (so it'll have all the same data, but a different name, and it'll appear in dashboard.convex.dev when the user logs in there).

Under the hood, this is
* creating a project
* calling `/local_deployment/start` on big brain to register the deployment as part of the project and get an instance name + admin key
* copy the contents of `tryitout-foo` to `local-sarahs_team-foo`
* update the config file of in `local-sarahs_team-foo`

There's an interactive flow under `npx convex login --link-deployments` if the developer wants to link these one by one -- we prompt this if the developer has more "tryitout" deployments than projects available on their team.

GitOrigin-RevId: 86983bea22755f5636d0f7931964506af4af0021