Add funnel tracking for no-account onboarding flow (#35884)
We want to roughly know how many people we have in the various stages of our funnel, and with a new experience where the user can try running Convex locally before signing up, we need to add some custom events.

Each device that tries out Convex without first making an account gets assigned a UUID. We record when a deployment is created (via the self hosted backend beacon), when the dashboard is visited, and when the user logs in.

GitOrigin-RevId: 38b6c931d663ba0e8410dfd06658650f26597b3f