Update messaging in CLI for initial flow (#35961)
The main change is adding a prompt on first use for whether to start without an account (default) or to create an account. It's gated behind the `CONVEX_TRY_IT_OUT` env var.

I also printed a link to the dashboard in one of the first messages in the dev loop (moved around code to make that happen).

Also made it so that the anonymous / self-hosted dashboard can accept a hint for the deployment name to select as a query param

GitOrigin-RevId: 2382d48deb29ae2c08f3216483f3bdc290ab859b