[Cron jobs] Fix behavior on cron job change (#35872)
We were seeing behavior where if a cron job was updated, we would calculate the next time to run it after the last run, which could be in the past. This could lead to immediately running a cron job unintentionally.

Now, the behavior is that we calculate the next time to run the cron job that matches the cron definition that is after right now. This feels more correct.

I also found some test cases that were incorrect because we were calculating the next time to run a job based on the fact that the last time a cron ran was in the future. This didn't make sense so I deleted it.

GitOrigin-RevId: d3f50d53af9f3eb6d58c530db14c97a717e724d0