Fivetran support destination connector schemaless (#35959)
Support fivetran destination connector schemaless.
This PR adds the indexes, but not the hidden tables.
Has get_schema return NotFound in schemaless mode.

Seems to work with the destination connector tester.

If a schema is provided, the old behavior continues.

Haven't bothered to clean up all the error messages/suggestions yet.
Want to see if we can push this out to play with it, and then
loop back to polishing error messages.

GitOrigin-RevId: b071cf1ecb5aac96b94721c1f21cc3072f7f2938