Fix crypto.subtle.sign for Ed25519 (#35993)
sign() was failing for Ed25519 keys because it was trying to get the internal key data
using `KEY_STORE.get(key).data`, which is incorrect because for Ed25519 keys
KEY_STORE stores the data directly - not behind an object.

This PR fixes that and adds a test exercising Ed25519 sign & verify.

GitOrigin-RevId: 8ead50fbd57d5a992dcba49f870f40cbbef9e3e8