Fix self-hosted dashboard UI when working on a cloud deployment (#35837)
On the project settings page, we disable a few features when using the self-hosted dashboard.
But if someone is working on a cloud deployment, we should redirect them to the cloud dashboard
if they want to use these features. This PR does that.

Note that this only detects cloud deployments if they are not using a custom domain. I think this is
fine for now.

GitOrigin-RevId: 49347a8a3e66f29c792e29dcb887a3592d68e0a4