Remove OCCs from push conflicting with cron jobs (#36219)
I reproed the /push_config OCC with an every-second cron job. Confirmed that this fixes!
This ensures that /push_config does not add the `next_run` ts into the readset in the case of a high-frequency cron job changing cronspec.

It also means that if you make changes unrelated to a high-frequency cron, the `next_run` won't get added to the read set.

Overall, this will reduce OCC to /push_config and improve availability on push!

GitOrigin-RevId: f13d64d3d322614d6e91a54cb7b50eb714474f87