Improve Sentry integration (#36485)
Because we've been using Sentry for Chef, I noticed the grouping in Sentry is pretty bad. This is because we use the `type` field to contain the error message (`type` in examples seems to be things like `ValueError` as opposed to long message), and we stringify the stack trace and stick it in the `value` field.

Instead, let's stick the error message in `value`, and use the `stacktrace` field to contain the stacktrace, and then sentry can group on stacktrace. I also have a story that sentry tries to group on `type` pretty aggressively, so not using `type` for something high cardinality will help.

Since this is technically a breaking change, I'm putting a little warning + dropdown in the dashboard to switch to the new format (we did something similar with log events).

I think Sentry is used to handling source mapping itself, but all of our `JsError`s already have been source mapped, so there's some weirdness around how these get displayed (sometimes there's a "could not symbolicate" warning in the Sentry UI), but this is still strictly better than just including the stacktrace as a string.

Oh also, made it so we include `ConvexError` and a custom context, because I think previously any structured data you passed along got lost

GitOrigin-RevId: 32ca17192a62282dd8c6cf4cae69fc43bc2caf41