design-system: support "all" selected state for MultiSelectCombobox (#36699)
When we use MultiSelectCombobox, we currently have to maintain the list of selected values all the time, even when the intent is to just select them all (the default state in all current scenarios). If the list of available values change, we would need to have logic to keep track of which values need to be added when all values are selected.

Instead, have MultiSelectCombobox and it's instantiating components be aware of an "all selected" state that means all values should be selected.

GitOrigin-RevId: b5d22ec4fbff31269b01a5a7414080d164f8fb40