Implement AES-GCM in isolate SubtleCrypto (#36734)
Only supports AES-128-GCM and AES-256-GCM with 96 bit nonces and 128 bit tags
(I'm using `ring` for crypto which makes it impossible to support anything else).

The other spec-defined algorithms (AES-CTR, AES-CBC, RSA-OAEP) aren't great and I
don't want to encourage their use for now. And we don't have a great provider for those.

It's on the user to ensure that nonces never get reused.

GitOrigin-RevId: c3d3f1e22be0dade13b6348fccaf2d20f5f116e2