docs: Fix title hierarchy in the ESLint page (#37149)
The title hierarchy of the ESLint docs page incorrectly used first-level titles (`h1`/`#`), which incorrectly placed these titles at the same level as the real first-level title of the page (“ESLint rules”). This also caused an issue in Docusaurus v3 where the actual title of the page wouldn’t be displayed because it would think we’re trying to replace the default `h1` by a custom  one.

GitOrigin-RevId: 785a7dde0325760c6e616253bdd34f875d72b2a1