docs: remove `/docs` from absolute Markdown links (#37132)
As indicated by https://docusaurus.io/docs/markdown-features/links, absolute file paths that are relaive to the site directory (i.e. `/npm-packages/docs/`) are not recommended, and it is recommended instead to use paths that are relative to the content root (`/npm-packages/docs/docs/`) instead.

I tested manually that links still work with this change.

GitOrigin-RevId: fd1738e467b4f7fc075fc9a1525e8db77da73ca8