Streaming export: stop using the legacy cursor format for list_snapshot (#37219)
This fixes a bug where streaming export would fail with an error message (_cannot find table with id 10001_) or never complete when exporting instances with components.

### What was happening?

`list_snapshot` uses a cursor that is essentially the last document that was seen. Its table number is ambiguous without knowing the component of the document. When we added component support for streaming export, we added a new cursor format that contains a tablet ID for disambiguation. However, the part of the code that sends these cursors to the connector was still using the legacy format. This would cause cursors to always be interpreted as being in the root component.

### Deployment
Even though this PR technically introduces an API breaking change (because it drops support for the legacy cursor format), the worst scenario that can happen is the conductor deployment happening between several `list_snapshot` calls. In this case, the user can restart the operation and it will work. The connectors simply treat `cursor` as an arbitrary string and don’t need updates. Also, only new Fivetran/Airbyte exports (or full resyncs of them) call `list_snapshot`.

### Testing
- I tested manually the new behavior by hitting the HTTP endpoints and following the cursors I received.
- I also added a new `application` unit test. This test used to fail when the cursor being passed didn’t contain the tablet ID, which is the previous behavior we had when using the legacy format. The test used to fail with an invalid table number error or infinite loop depending on the order of the tablet IDs, which matches the behavior we had in production repros of this bug.
- This code path is also tested by a smoke test, that I’ve fixed in #37218

GitOrigin-RevId: 0f0c0ea511d5cb07d0a9f4615305f3b07ac2312b