Optimize CountedShape::remove for unions (#37403)
The existing code builds up an entire new union from an empty UnionBuilder,
which is quite inefficient as each call to `UnionBuilder::push` needs to recheck
invariants and possibly collapse variants. However, since almost all variants
came from a previous valid UnionShape, we can skip that work except for the one
variant that was actually modified.

This code is on the critical path of the committer and is rather expensive in
practice.

GitOrigin-RevId: f8e85d7cc0601d7338f4cbeaa1f8aaf4e4a5d1e7