Fix `defineTable({ ...v.object({}) })` handling (#37561)
When writing code like `defineTable({ ...v.object({}) })`, the validator’s attributes are copied but not getters, including `json` which is used during the table definition export.

This means that `export` would silently fail, as it did for someone on Discord (https://discord.com/channels/1019350475847499849/1019350478817079338/1374853725491036270). Other schema definition errors are handled by the backend, but this one isn’t since `document_type: None` is a valid attribute value in `TableDefinitionJson`.

`export` in `TableDefinition` is the only `export` method where we’re calling a JavaScript getter, so other code paths should be fine.

GitOrigin-RevId: 392fa3b7cbd1939dae8349e01dcde8ea2e29c079