Fix assertNotBrowser false negatives (#37912)
`__convexAllowFunctionsInBrowser` incorrectly _enables_ the assertion instead of disabling it. To avoid making a breaking change for now, we log an error instead of throwing for now, but we will start throwing runtime errors again at some point in the future. The error message asks users to contact the Convex support so that we can fix false positives if necessary, but users can still disable the warning using `window.__convexAllowFunctionsInBrowser`.

GitOrigin-RevId: ba259c906bc5eafb60b6dbae13e5f69708c10f91