Refactor `DocumentTokens` so that it doesn’t hold the entire document (#38036)
This refactors `query.rs` so that `DocumentTokens` doesn’t need to hold the entire document but only the value for a particular field. This will be helpful for future read-set intersection improvements (ENG-4475) where we we only have the value of the search field instead of the entire document.

Summary of changes:

- `DocumentTokens` becomes `ValueTokens`
  - It is constructed from a `ConvexString` instead of referencing a `PackedDocument`
- `FieldTokens` is merged into `ValueTokens`
- `matching_values` in `SearchTermTries` is modified so that it works over a single field instead of the entire document, and its usages are updated to still get the same behavior as before. I also moved it to `Tries`.

I added tests to verify that the behavior of the code isn’t changed.

GitOrigin-RevId: e60cd5dcc14de155b1a5fbc417a911f41216377d