Don't crash when optimistic updates exist on queries for which no server result exists yet (#38020)
Fix a bug with setting optimistic updates for recently- or not-yet-subscribed-to queries.

Previously the client threw an error and aborted handling a server update with "Cannot read properties of undefined (reading 'result')" in `BaseConvexClient.notifyOnQueryResultChanges` (among other more obfuscated frames) when the client set an optimistic update for a query for which it did not yet have a value from the server.

GitOrigin-RevId: 1fb59b02d5f14750a4d8bdf8c1bb259c4dfdccb5