Rename overlaps → overlaps_document (#38134)
This is helpful for ENG-4475 which will introduce variants of these methods that take a `&DocumentIndexKeys` instead of a `&PackedDocument`.

I also made the `overlaps` method in `ReadSet` that doesn’t take a buffer test-only since real production code should always provide a reusable buffer for efficiency. (This method was already only used in tests).

GitOrigin-RevId: 4c431f2b77e1c86ea5cfb51c85a98b7d77416812