Add data structures for storing keys in write log (#38131)
This adds `DocumentIndexKeys`, which will be used in the future to store the index keys of an object so that we can store only the keys in the write log. This will save memory and make read-set intersection more CPU efficient, as the keys won’t need to be recomputed every time.

This PR doesn’t modify the write log yet but adds some code that will be helpful later.

GitOrigin-RevId: c6cfd54bdfb0893f30fdad6fb97cabb3b0111672