Make sure that spending limit errors are shown on form submission (#38182)
This PR fixes the error handling of the spending limit form in the situations where the user enabled a checkbox but didn’t write a value.

To do so, I changed both the spending limits form and the upgrade form so that their submit buttons are not disabled when there is a Formik error: instead, when the button is clicked Formik sets both fields as “touched”, which will force errors to be shown.

To avoid an issue where Formik wouldn’t set fields with a value of `undefined`, which is the value we had for unset spending limit fields when the checkbox is enabled, I changed the value from `undefined` to `""`. I moved the code that converts the `SpendingLimitValue` in both forms to a function and added tests.

GitOrigin-RevId: 58c539a20ad55dab1ea2616a352546a4ed953845