convex-backend PR 108: print tablename in unique() error message (#38187)
Improves usefulness of the error thrown when `.unique()` returns more than 1 result by now also printing the table name.

Improves quality of life: currently, it can be cumbersome to identify the actual table in which these non-unique elements belong.

This solution is a bit work-around-y, but it's the only solution I could come up with. Let me know if there is a better way to achieve this!

----

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.

Co-authored-by: Jan-Georges Jersild Balin <jgbalin@gmail.com>
GitOrigin-RevId: 3b5e5bc6888326f8a6d615c3272cdf66aa285987