onServerDisconnectError callback (#37850)
Stopgap client callback to report unusual server disconnect errors.

This would be an ugly API to be stuck with for the long term because it leaks details that are irrelevant, but customers have requested a way to know on their users' clients that the connection is down so we're getting this out quickly.

The content of these messages, which messages trigger the callback, and the existence of the callback are all subject to change.

GitOrigin-RevId: 86c37e76ee0517f1dee54273ec6ac129021371d2