Make SearchCompactor backoff configurable (#38417)
Makes knobs for setting max backoff for search index compaction and flushing. Flushing is critical and should have lower max backoff, but compaction is less important, and if it is consistently failing, we stop sending frequent requests. The impact of this change is increasing max backoff from 30s to 10min for compaction. We could split out these knobs for text and vector indexes, but the principle is the same for both so to avoid knob bloat, let's just use the same knobs.

GitOrigin-RevId: 7fdc47f74b2e3373e01298f0031b082bfd271d90