[1/3] Add a new API for system queries (#38452)
This adds a new API `tx.system_query()` which lets Rust code query `SystemTable`s
in an ergonomic and more-type-safe fashion, using the `<T as SystemTable>::Metadata`
associated type.

The API is also designed so that in-memory indexes can (in a followup PR) cache the
_parsed_ values of system documents, which will save a lot of CPU re-parsing in e.g.
ModulesModel, UdfConfigModel, etc.

In general this avoids a lot of extra checks and heap allocations that are incurred using
the `ResolvedQuery` code path.

This does incur a moderate amount of new "interesting" code, mainly around constructing
index intervals; I've included a fair few tests for that.

GitOrigin-RevId: 2715c151dbbd17c2dfde57b5d07c15b1398f3ba7