[3/3] Cache parsed documents across transactions (#38454)
This adds a cache to `BackendInMemoryIndexes` that holds a per-document `Arc<dyn Any>`,
which is populated with the correctly-typed `Arc<ParsedDocument<T::Metadata>>` when
queried via a system query. This avoids the overhead of re-parsing system documents each
time they are queried.

If the same document is queried twice with _different_ metadata types, this will panic in debug
and print a warning in release. This should never happen as long as SystemTable implementations
never point at the same table. For migrations I've added a new const to SystemTable to turn off
typed caching.

The double-Arc is unfortunate but I don't have a great solution to that without pulling in heavy machinery.

GitOrigin-RevId: ff0d1211229b30a286f6c721e7e4fd9e7ae52926