Skip sending search query results while bootstrapping (#38690)
This PR changes the behavior of search queries when search indexes are bootstrapping. Instead of returning `SearchIndexesUnavailable` error (an internal server error that causes the websocket to close), we skip returning any results. This means already-subscribed search queries will show stale results until indexes are bootstrapped, which is a significant change from current behavior which guarantees a consistent view of data (e.g. a mutation that adds a doc that would match the search query won't immediately show up in the results).

GitOrigin-RevId: bdaefb20e262bb1718f31ef50439121929c7947f