Add a `table` argument to UDF syscalls (#38673)
This PR adds a new `table` argument to the following UDF syscalls:

- `1.0/get` (`db.get`)
- `1.0/shallowMerge` (`db.patch`)
- `1.0/replace` (`db.replace`)
- `1.0/remove` (`db.delete`)

The new argument is optional, which makes the change backwards compatible (i.e. existing UDF bundles can still be used). When `table` is defined, we only use it to verify that its value matches the ID that was provided, and throw an error when it is not the case.

This new API is used in the new user-facing APIs added by #38505. While this PR has no tests, the tests from #38505 verify that the behavior encountered when calling these syscalls with the wrong table name is correct.

GitOrigin-RevId: 847ffe892204dca0faf940ad4eb1eaafc7b2218b