Move index_backfills table to bootstrap tables to avoid circular dependency (#38772)
We need index workers to be able to write to `_index_backfills` table, but index workers live in the `database` crate and `_index_backfills` is in `model` crate. It would be nice if we could move index workers out into their own crate, but this is a big project that involves moving almost all of the database test code as well. So for now we're putting the `_index_backfills` table in the bootstrap tables

GitOrigin-RevId: 374956d32df4a88d7b7d64e7bde15ecc269bcf4a