Store search tokens in the write log (#38302)
This pull request changes `SearchIndexKeyValue` so that it stores a `SearchIndexKeyValue` (which is essentially a `Vec<String>`) instead of a `ConvexString`. This avoids needing to tokenize the values of search documents during read set intersection (→ –75% improvement on the `is_stale` benchmarks involving search indexes).

`SearchIndexKeyValue` stores the tokens efficiently, using `CompactString` + a boxed slice. This works since we only iterate on all of its elements, so we don’t need to efficiently determine if a token is part of it once it’s created.

GitOrigin-RevId: bbf3eaa2fe3060aaf7104e1cfa502dc4278897bc