Add a new `version` service to know the current NPM + Cursor rules version (#38740)
This PR adds a new service that will be used by the CLI to know about the latest version of the `convex` NPM package and the Convex cursor rules.

It exposes two HTTP endpoints:
- `/v1/version`, which returns something like `{"message": "A new version of Convex is available…", "cursorRulesHash":"12f7147d03dcb815ab7216942e42b4c4d15b21c1f5d6873f179c47e9aac35f97"}`
- `/v1/cursor_rules`, which is the latest version of the Convex Cursor rules.

The NPM package version is obtained from the NPM HTTP API. The Cursor rules are downloaded from the convex-evals repo using the GitHub API, using the same code as used by [create-convex](https://github.com/get-convex/templates/blob/main/create-convex/src/cursorRules.ts).

To allow the endpoints to return responses quickly and not do too many API calls, the results are cached in the Convex DB for an hour. When a response is stale, we respond using the stale response and then schedule an update. If there is no data cached, we fetch the necessary data synchronously.

GitOrigin-RevId: 7fd6d4db9811647f75d1eaf1b752e2d621a38eff