Load in-memory tables at startup once, not once-per-index (#39022)
This aims to reduce DB traffic on the underlying persistence during startup,
at the cost of recalculating some index keys. I believe that is cheap enough
that the extra CPU is already offsetted by skipping the extra index scans.
This also saves a tiny bit of memory as the PackedDocuments can be shared
between index maps.

GitOrigin-RevId: 467b477fb12c067a6d311170705f6749e2b4ecb0