Write index backfill progress metadata (#39176)
This PR adds code that writes index backfill progress. I simplified `IndexBackfillMetadata` to only track documents because size is hard to get right. `totalDocs` is optional so we can track progress as table summaries are bootstrapping and fill in the approximate size of the table later. We will be careful in the UI to not represent over 100% completion.

GitOrigin-RevId: a05320b13a19c55fa4c207509d26e4f4f064e988