Replace IntervalMap with a BBST-based implementation (#39326)
This replaces the naive IntervalMap implementation (which queries each
subscription against each index key separately) with a treap-based data
structure that supports querying an index key in O((k+1) log n) time, where k is the
number of _matching_ subscriptions and n is the total number of subscribed
intervals. The net effect is that long-lived subscriptions get much cheaper.

GitOrigin-RevId: f39ee382b2ca4441b71b64715c52d27a2556bcc5