Allow deploy keys to perform actions that required a member previously (#39715)
With the launch of the Convex platform API, we will want to share endpoints used for managing deployments, and have them be accessible via deploy keys (service tokens). Previously, we checked for a member identity to allow performing certain actions like deleting tables or creating an integration, or retrieving shapes. Now, deploy keys will be able to do these things

GitOrigin-RevId: 251db8d9a9a338af4419965d94da9f4bb798ff3a