convex-backend PR 182: Update using-cursor.mdx (#39899)
Hey!

Noticed the documentation for adding Convex docs to cursor is a bit outdated and do not properly reflect the current state of how it should be done.

1. The way to add to docs is now under a different menu `Cursor settings` > `Indexing & Docs` > `Docs`

<img width="981" height="453" alt="image" src="https://github.com/user-attachments/assets/77d04f93-037d-41af-919c-355baf65f8a9" />

2. It seems that you guys redirect `https://docs.convex.dev/` to `https://docs.convex.dev/home` and therefore when adding the url currently showcased in the docs, cursor its not able to index the documentation. The proposed url makes cursor able to correctly index the 278 pages of docs.

----

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.

Co-authored-by: Ricardo Alvarez-Correa <ricardo@binah.co>
GitOrigin-RevId: 0962921dc177de563b33418899cd131f51fde15e