Enable backends to use fine-grained streaming export selection (#37364)
This PR modifies the arguments for table selection of `list_snapshot` and `document_deltas`: instead of taking `table_name` and `component`, they now take a `SelectionArg`, which allows for fine-grained selection of the data to stream.

`SelectionArg` is designed to be fully backwards-compatible: the previously existing `table_name` and `component` arguments are still supported as enum variants of `SelectionArg`. In addition to that, `SelectionArg::Exact` allows the API users to use the new selection APIs introduced by #37346.

This PR also changes the relevant code in the `database` and `conductor` crates so that the precise selection is taken into account. Two new tests in the `streaming_export_tests.rs` verify that selection works as intended.

GitOrigin-RevId: 05891a2a6917be45f712c6c590cb1544b7d579e3