relaxing the log entry validator strictness (#40019)
While recording a video on the mcp logs tool I started getting errors when trying to return the logs. I think for some reason some fields were missing from the log entries and thus the output validator was rejecting it.

I dont think we actually need a strict output validator in this case as it is just an MCP tool so the agent is going to be able to interpret the data anyways, we dont need to first validate it.

GitOrigin-RevId: e085061f044a1dacebc957d369bdfe59d9e1b0e0