[Node Migration] Add version to `_source_packages` (#40102)
Adds a field on the `_source_packages` system table that specifies which node version a customer has specified. This field is nullable and backwards compatible, so it doesn't need a backfill.

I also added this field as optional on push endpoints. Next, we will have to start sending up this value from the CLI.

GitOrigin-RevId: acff5dfc9b4d027a56f32fc2e778fe820ac31edc