convex-backend PR 196: Fix: Respect `DISABLE_BEACON` env var in backend (#40167)
This change addresses issue #194  by making the backend binary directly obey the DISABLE_BEACON environment variable, in addition to the existing --disable-beacon flag.

Previously, the environment variable was only handled by the Docker Compose setup, which was confusing for users not using Docker.

This PR also clarifies the startup message to reflect that either the environment variable or the command-line flag can be used to disable telemetry.

----

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.

Co-authored-by: Shoaib <starvoltz173@gmail.com>
GitOrigin-RevId: 16af2efdb891547324020847e5359bfa53d46911