Make infinite or NaN float literals bad requests (#40440)
Serializing infinite or NaN values in literals fails (I guess they are not valid JSON values), so we should just make this a bad request instead of internal server error so users can fix it. It doesn't look like we call `serde_json::Number::from_f64` anywhere else that we need to fix (only other callsite is for display and we convert to a string if it is `None`).

GitOrigin-RevId: 982ff69d541c6783ee483745672e38173c032843